Change the gamma override value config to a double

+++ src/main/java/fi/dy/masa/tweakeroo/config/Callbacks.java
@@ -147,7 +147,7 @@ public class Callbacks
             // If the feature is enabled on game launch, apply it here
             if (feature.getBooleanValue())
             {
-                this.mc.options.gamma = Configs.Generic.GAMMA_OVERRIDE_VALUE.getIntegerValue();
+                this.mc.options.gamma = Configs.Generic.GAMMA_OVERRIDE_VALUE.getDoubleValue();
             }
         }
 
@@ -157,7 +157,7 @@ public class Callbacks
             if (config.getBooleanValue())
             {
                 Configs.Internal.GAMMA_VALUE_ORIGINAL.setDoubleValue(this.mc.options.gamma);
-                this.mc.options.gamma = Configs.Generic.GAMMA_OVERRIDE_VALUE.getIntegerValue();
+                this.mc.options.gamma = Configs.Generic.GAMMA_OVERRIDE_VALUE.getDoubleValue();
             }
             else
             {
